rename movie components
